Fix #3205: VE crash for objects with length exceeding 32-bit

+++ src/Host/API/Impl/Definitions/IRObjectInformation.cs
@@ -16,11 +16,11 @@ namespace Microsoft.R.Host.Client {
         /// <summary>
         /// Object length (such an number of elements in a list)
         /// </summary>
-        int Length { get; }
+        long Length { get; }
 
         /// <summary>
         /// Number of dimensions in the object.
         /// </summary>
-        IReadOnlyList<int> Dim { get; }
+        IReadOnlyList<long> Dim { get; }
     }
 }
